Pattern bugfix

+++ mcu/src/sequences.cpp
@@ -16,7 +16,7 @@ void fill_and_show(uint32_t c, Adafruit_NeoPixel *strip) {
 // fills the strip by repeating the colors provided in array c
 // len is the number of colors in array c
 // number of provided colors can't exceed number of unique pixels
-void fill_pattern(uint32_t *c, uint8_t len, Adafruit_NeoPixel *strip) {
+void fill_pattern(uint32_t *c, uint16_t len, Adafruit_NeoPixel *strip) {
     for (int i=0; i<strip->numPixels(); i++) {
         strip->setPixelColor(i, c[i % len]);
     }
